Create server action to get lesson plans
